Vote repo structure (#16)
* Move most repo methods into base interfaces

* Remove methods from IIssueRepo; Inherit from base class

* Change Find from nullable int to regular int

* Explictly cast int? to int in controller

* Add IVoteRepo to project; update csproj

* Setup VoteRepo for DB Access; Alter DB Context to allow access to votes

* Rename Issue Tests to align with project standard

* Change GetAllVotesForIssue to take an int Id

* Inject dependency of db context to enable easy mocking

* Make dbSets virtual to allow mocking

* Refactor test names

* Allow AddVote to be mocked

* Add Vote Repo to project

* Write Repo tests; bring Moq into project

+++ LobbyingMadeSimple/Controllers/IssuesController.cs
@@ -37,7 +37,7 @@ namespace LobbyingMadeSimple.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Issue issue = _repo.Find(id);
+            Issue issue = _repo.Find((int) id);
 
             if (issue == null)
             {
@@ -80,7 +80,7 @@ namespace LobbyingMadeSimple.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Issue issue = _repo.Find(id);
+            Issue issue = _repo.Find((int) id);
 
             if (issue == null)
             {
@@ -117,7 +117,7 @@ namespace LobbyingMadeSimple.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Issue issue = _repo.Find(id);
+            Issue issue = _repo.Find((int) id);
 
             if (issue == null)
             {
